Make State immutable
Refactor the State class so it is immutable. All setter functions now
return a new State instance.

This change is less important than making the Tree immutable but I'm
hoping it will have a positive impact and will make debugging and unit
testing easier.

The biggest consequence is that you have to remember to update your
reference to use the new version of state whenever there are changes.

Ex.

This is right:

```
state = state.setFoo()
```

This is wrong (does nothing):

```
state.setFoo()
```