Create Tree and AgsService modules
Refactors the State module by splitting it up into 3 distinct modules:

* State - Represents plugin saved state; contains simple getter/setters

* Tree - Collection of layers; has filter/find functions as well as all
the data necessary to render the tree in the UI

* AgsService - Responsible for all ajax requests

The purpose of this change is to create several smaller units of code that
can be easily tested.

Other changes:

* LayerNode should now be the single source of truth for all layer data.
As a consequence, the tree must be recreated every time there is a state
change or map service data is fetched.

* Events have been removed. These didn't seem necessary anymore and
should make unit testing simpler.