Show info button for parent nodes
Adjust the style and layer node template so that the info button is shown
for parent nodes. Some layer group parents have their own descriptions.

Update the sample configuration file with a service that has a layer group
with a description and to show that top-level layers can also have a
description.

Refs #50