Render transparency correctly
The default image format of png8 would blend transparent images
against a solid white background, and render that blended image
which would not be actually transparent on the front-end. png32
on the other hand renders transparent images correctly, thus is
the preferred image format.

Unfortunately, it needs to be set before every call, and cannot
be used during layer initialization. We must be careful of this
potentially overloading the server if many layers are visible.