Narrow scope for menu item event binding
The menu items were registered with events bound to `body` which, if
multiple instances of regional-planning were included, would trigger
handlers in incorrect instances, resulting in errors. The previous
implementation of the menu drop down probably necessitated the placement
originally, but it has since been refactored to live in the DOM tree under
the plugin container, and not on body.  The event is registered more
narrowly on the container now.